MDEV-26247 [wip] Re-implement spider gbh query rewrite of tables
Spider GBH's query rewrite of table joins is overly complex and
error-prone. In this commit, we explore replacing it with
something closer to what dbug_print() (more specifically,
print_join()) does, but catered to spider.

We make it handle const tables by rewriting them as (select 1). This
fixes the main issue in MDEV-26247.

We make spider_create_group_by_handler() go through append_join too
and get rid of append_from_and_tables() and friends.

There is one tricky issue remaining:

By the time the query reaches the spider gbh, the optimizer may
have modified the query so that it is no longer valid. See the test
tmppp.test

However, this is not a regression, as the test fails before this
change.