MDEV-28561 Assertion failed: !pfs->m_idle || (state == PSI_SOCKET_STATE_ACTIVE)
The error is specific to threadpool.It would attempt to set socket state
to idle, when it was already idle, and this into a DBUG assertion.

This happened when there was still some unread compressed data buffered
on connection. This can easily happen when client sends STMT_CLOSE followed
by any other command, thus testing with ps-protocol is recommended.